ft(table):integrate the requests table in the backend
Created the controller and router setup to allow allow users retrieve travel requests.

[Finishes #167727463]
