+++ .env.example
@@ -7,17 +7,28 @@
 # Please take note that the dummy values below will not work.
 # They are just to show you what yours shoud look like.
 
-DB_URI_DEVELOPMENT="https://devwebsite.com"
+NODE_ENV=development
+
+DB_DIALECT=postgres
+
+DB_STRING_DEVELOPMENT=dev://dev:devpasswd:5555/
+DB_HOST_DEVELOPMENT=localhost
+DB_NAME_DEVELOPMENT=db_dev
 DB_USER_DEVELOPMENT=devUser
 DB_PASSWORD_DEVELOPMENT=devPassword
 
-DB_URI_STAGING="https://stagingWebsite.com"
+DB_STRING_STAGING=staging://staginguser:stagingpasswd:5432/
+DB_HOST_STAGING=127.0.0.1
+DB_NAME_STAGING=db_staging
 DB_USER_STAGING=stagingUser
 DB_PASSWORD_STAGING=stagingUser
 
-DB_URI_TEST="https://prodwebsite.com"
-DB_USER_TEST=prodUser
-DB_PASSWORD_TEST=prodPassword
+DB_STRING_PRODUCTION="prod://produser:prodpasswd:3306/"
+DB_HOST_PRODUCTION=localhost
+DB_NAME_PRODUCTION=db_prod
+DB_USER_PRODUCTION=stagingUser
+DB_PASSWORD_PRODUCTION=stagingUser
+
 
 .................
 .................
