+++ .eslintignore
@@ -1,0 +1,9 @@
+src/test
+.gitignore
+HISTORY.txt
+README.txt
+node_modules/
+.env
+app.js
+.idea/
+dist/
