+++ .eslintrc.json
@@ -7,17 +7,21 @@
         "mocha": true,
         "commonjs": true
     },
+    "parser": "babel-eslint",
     "parserOptions": {
         "ecmaVersion": 6,
         "sourceType": "module",
+        "allowImportExportEverywhere": true,
         "ecmaFeatures": {
             "jsx": true
         }
     },
-    "parser": "babel-eslint",
     "plugins": [
         "flowtype"
     ],
+    "globals": {
+        "_": false
+    },
     "rules": {
         "one-var": 0,
         "camelcase": 0,
