+++ .mocharc.json
@@ -1,3 +1,3 @@
 {
-    "spec": "src/test"
+  "spec": "src/test"
 }
