+++ package.json
@@ -15,7 +15,7 @@
         "docker:rebuild": "docker-compose up --rebuild",
         "docker:destroy": "docker-compose down -v",
         "lint": "eslint ./src --fix",
-        "nyc": "nyc mocha -r esm --cache false --timeout 3000 --exit",
+        "nyc": "nyc mocha -r esm --cache false --timeout 30000 --exit",
         "db:migrate": "sequelize db:migrate",
         "db:seed": "sequelize db:seed:all --debug",
         "undo:migration": "sequelize db:migrate:undo:all",
@@ -26,7 +26,9 @@
     },
     "nyc": {
         "exclude": [
-            "src/index.js"
+            "src/index.js",
+            "src/db/*",
+            "src/routes/api/index.js"
         ]
     },
     "author": "Andela Simulations Programme",
@@ -54,6 +56,7 @@
         "passport-local": "^1.0.0",
         "pg": "^7.12.1",
         "pg-hstore": "^2.3.3",
+        "proxyquire": "^2.1.3",
         "pug": "^2.0.4",
         "regenerator-runtime": "^0.13.3",
         "request": "^2.87.0",
@@ -80,7 +83,11 @@
         "eslint-plugin-flowtype": "^4.2.0",
         "eslint-plugin-import": "^2.18.2",
         "mocha": "^6.1.4",
+        "nock": "^10.0.6",
         "nodemon": "^1.19.1",
-        "nyc": "^14.1.1"
+        "nyc": "^14.1.1",
+        "sinon": "^7.4.1",
+        "sinon-chai": "^3.3.0",
+        "supertest": "^4.0.2"
     }
 }
