+++ src/routes/api/auth.router.js
@@ -1,13 +1,12 @@
 import express from 'express';
-
 import UserController from '../../controllers/user.controller';
 import UserMiddleware from '../../middlewares/user.middleware';
 
 const router = express.Router();
 
 /* Users Routes Here */
-router.post('/signup', UserController.signup);
-router.post('/signin', ...UserMiddleware.validateSigninFields(),
+router.post('/auth/signup', UserController.signup);
+router.post('/auth/signin', ...UserMiddleware.validateSigninFields(),
   UserController.signin);
 
 export default router;
