+++ src/routes/api/index.router.js
@@ -1,0 +1,8 @@
+import express from 'express';
+const router = express.Router();
+
+router.get('/', (req, res) => {
+    res.status(200).json({ Message: 'Welcome! This is the NorthStar Barefoot Nomad homepage' });
+});
+
+export default router;
